Restruct supervised learning location
