+++ bark_ml/tests/capability_GNN_actor/BUILD
