--- supervised_learning/config.py
@@ -1,1 +1,0 @@
-../config.py
